Renaming hare2 -> hare3
